refactor: extract entry utils to utils package
