+++ apps/frontend/src/components/questionnaire/calendar/EntryCalendar.tsx
@@ -3,14 +3,13 @@ import interactionPlugin from "@fullcalendar/interaction";
 import timeGridPlugin from "@fullcalendar/timegrid";
 import { DateSelectArg, EventChangeArg, EventInput } from "@fullcalendar/core";
 import { Button, Modal, useDisclosure, useMantineTheme } from "@quassel/ui";
-import { getDateFromTimeAndWeekday, getTime, isSame } from "@quassel/utils";
+import { GapsPerDay, getDateFromTimeAndWeekday, getTime, groupByWeekday, isSame } from "@quassel/utils";
 import { QuestionnaireEntry } from "./QuestionnaireEntry";
 import { components } from "../../../api.gen";
 import { EntityForm, EntryFormValues } from "./EntryForm";
 import { useEffect, useState } from "react";
 import { format, i18n } from "../../../stores/i18n";
 import { useStore } from "@nanostores/react";
-import { GapsPerDay, groupByWeekday } from "../../../utils/entry";
 import { EventImpl } from "@fullcalendar/core/internal";
 import styles from "./EntryCalendar.module.css";
