+++ libs/utils/package.json
@@ -21,7 +21,8 @@
     "build": "vite build --minify",
     "build:dev": "vite build --watch --sourcemap",
     "lint": "eslint .",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "dependencies": {
     "dayjs": "^1.11.13"
@@ -30,6 +31,7 @@
     "@eslint/js": "^9.19.0",
     "typescript": "^5.7.3",
     "vite": "^6.0.11",
-    "vite-plugin-dts": "^4.5.0"
+    "vite-plugin-dts": "^4.5.0",
+    "vitest": "^3.0.3"
   }
 }
