+++ libs/utils/src/index.ts
@@ -1,1 +1,3 @@
 export { formatDate, getTime, getDateFromTimeAndWeekday, getNext, isSameOrAfter, isSame } from "./date";
+
+export { type Gap, type GapsPerDay, groupByWeekday, resolveGaps } from "./entry";
