+++ libs/utils/tsconfig.json
@@ -4,7 +4,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "lib": [
-      "ES2015",
+      "ES2024",
       "DOM"
     ],
     "module": "ESNext",
